project-board/src/main/java/com/practice/project/service/ArticleService.java
package com.practice.project.service;

import com.practice.project.domain.Article;
import com.practice.project.domain.UserAccount;

import com.practice.project.domain.type.SearchType;
import com.practice.project.dto.ArticleDto;
import com.practice.project.repository.ArticleRepository;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import javax.persistence.EntityNotFoundException;
import java.util.Arrays;
import java.util.List;
import java.util.Set;
import java.util.stream.Collectors;

@Slf4j
@RequiredArgsConstructor
@Transactional
@Service
public class ArticleService {
    private final ArticleRepository articleRepository;

    @Transactional(readOnly = true)
    public Page<ArticleDto> searchArticles(SearchType searchType, String searchKeyworde) {
        return Page.empty();
    }



    @Transactional(readOnly = true)
    public ArticleDto getArticle(Long articleId) {
        return null;
    }

    public void saveArticle(ArticleDto dto) {

    }
}
