project-board/src/test/java/com/practice/project/service/ArticleServiceTest.java
package com.practice.project.service;

import com.practice.project.domain.Article;

import com.practice.project.domain.UserAccount;

import com.practice.project.domain.type.SearchType;
import com.practice.project.dto.ArticleDto;
import com.practice.project.repository.ArticleRepository;

import org.assertj.core.api.InstanceOfAssertFactories;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageImpl;
import org.springframework.data.domain.Pageable;
import org.springframework.test.util.ReflectionTestUtils;

import javax.persistence.EntityNotFoundException;
import java.time.LocalDateTime;
import java.util.HashSet;
import java.util.List;
import java.util.Optional;
import java.util.Set;
import java.util.stream.Collectors;

import static org.assertj.core.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.*;
@DisplayName("비즈니스 로직 - 게시글")
@ExtendWith(MockitoExtension.class)
public class ArticleServiceTest {
    @InjectMocks private ArticleService sut;

    @Mock private ArticleRepository articleRepository;

    @DisplayName("검색어 없이 게시글을 검색하면, 게시글 페이지를 반환한다.")
    @Test
    void givenNoSearchParameters_whenSearchingArticles_thenReturnsArticlePage() {
        // Given

        // When
        Page<ArticleDto> articles = sut.searchArticles(SearchType.TITLE, "test keyword");

        // Then
        assertThat(articles).isNotNull();
    }

    @DisplayName("게시글 정보를 입력하면, 게시글을 생성한다.")
    @Test
    void givenSearchParameters_whenSavingArticles_thenSaveArticlePage() {
        // Given
        given(articleRepository.save(any(Article.class))).willReturn(null);
        // When
        sut.saveArticle(ArticleDto.of(LocalDateTime.now(),"dev","title","content","hashtag"));

        // Then
        then(articleRepository).should().save(any(Article.class));
    }



}
